refactor: clarify session history builder
