+++ packages/server/src/tools/core/inspect-image.ts
@@ -15,8 +15,8 @@ import * as AbortRegistry from '@/lib/abort-registry.js';
 import { internalBus } from '@/lib/internal-bus.js';
 import * as Log from '@/lib/log.js';
 import { isServiceError } from '@/lib/service-result.js';
-import { buildCompactedHistory } from '@/llm/compaction.js';
 import type { ProviderCredentials } from '@/llm/provider/provider.js';
+import { buildSessionLlmMessages } from '@/llm/session-history.js';
 import { runStream } from '@/llm/stream/runner.js';
 import type { ToolContext } from '@/tools/runtime/runtime.js';
 
@@ -173,7 +173,7 @@ export function createInspectImageTool(context: ToolContext, deps: InspectImageT
         duration: null,
       });
 
-      const llmMessages = await buildCompactedHistory(childSessionId, {
+      const llmMessages = await buildSessionLlmMessages(childSessionId, {
         useBasePrompt: true,
         systemPrompt: null,
       });
